pyproject.toml
[project]
name = "vgquiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.2",
    "django-ninja>=1.4.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
